Moved metadata builders to Builders namespace
