+++ src/DWC_A/ArchiveWriter.cs
@@ -1,5 +1,4 @@
 ﻿using DwC_A.Builders;
-using DwC_A.Meta;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
