+++ src/Tests/ArchiveWriterTests.cs
@@ -1,6 +1,5 @@
 ﻿using DwC_A;
 using DwC_A.Builders;
-using DwC_A.Meta;
 using DwC_A.Terms;
 using DwC_A.Writers;
 using System;
