+++ src/Tests/CoreFileMetaDataBuilderTests.cs
@@ -1,4 +1,4 @@
-﻿using DwC_A.Meta;
+﻿using DwC_A.Builders;
 using DwC_A.Terms;
 using System.Text;
 using Xunit;
