+++ src/Tests/FieldMetaDataBuilderTests.cs
@@ -1,4 +1,5 @@
-﻿using DwC_A.Meta;
+﻿using DwC_A.Builders;
+using DwC_A.Meta;
 using DwC_A.Terms;
 using Xunit;
